fix some warnings inside CxxTest, second attempt

+++ lib/cxxtest/cxxtest/ValueTraits.cpp
@@ -36,7 +36,7 @@ char digitToChar(unsigned digit)
 const char *byteToHex(unsigned char byte)
 {
     static char asHex[3];
-    asHex[0] = digitToChar(byte >> 4u);
+    asHex[0] = digitToChar(unsigned(byte >> 4));
     asHex[1] = digitToChar(byte & 0x0Fu);
     asHex[2] = '\0';
     return asHex;
